Split public api code and internal code.
Code is now organized like this:
- Public Interfaces are now in api_interfaces.go
- Public API is now in api_public.go
- All internal stuff remains in api.go for now
