+++ api_interfaces.go
@@ -15,6 +15,16 @@ type CRUD interface {
 	Update(obj interface{}, req Request) error
 }
 
+// ContentMarshaler controls how requests from clients are unmarshaled
+// and responses from the server are marshaled. The content marshaler
+// is in charge of encoding and decoding data to and from a particular
+// format (e.g. JSON). The encoding and decoding processes follow the
+// rules of the standard encoding/json package.
+type ContentMarshaler interface {
+	Marshal(i interface{}) ([]byte, error)
+	Unmarshal(data []byte, i interface{}) error
+}
+
 // The PaginatedFindAll interface can be optionally implemented to fetch a subset of all records.
 // Pagination query parameters must be used to limit the result. Pagination URLs will automatically
 // be generated by the api. You can use a combination of the following 2 query parameters:
